Keep flash logic and boot modes in zsw_flash_loader.
